+++ libs/mops/src/thds/mops/pure/adls/output_fqn.py
@@ -1,7 +1,7 @@
 from thds.adls import AdlsFqn
 
-from ..core import types, uris
-from ..core.output_naming import pipeline_function_invocation_unique_key
+from ..core import uris
+from ..core.output_naming import invocation_output_uri
 
 
 def invocation_output_fqn(storage_root: uris.UriIsh = "", name: str = "") -> AdlsFqn:
@@ -14,17 +14,4 @@ def invocation_output_fqn(storage_root: uris.UriIsh = "", name: str = "") -> Adl
     <pipeline> <function mod/name  > <your name     > <args,kwargs hash                                   >
     nppes/2023/thds.nppes.intake:run/<name goes here>/CoastOilAsset.IVZ9KplQKlNgxQHav0jIMUS9p4Kbn3N481e0Uvs
     """
-    storage_root = storage_root or uris.ACTIVE_STORAGE_ROOT()
-    pf_fa = pipeline_function_invocation_unique_key()
-    if not pf_fa:
-        raise types.NotARunnerContext(
-            "`invocation_output_fqn` must be used in a `thds.mops.pure` runner context."
-        )
-    pipeline_function_key, function_arguments_key = pf_fa
-    return (
-        AdlsFqn.parse(str(storage_root))
-        / pipeline_function_key
-        / name
-        / "--".join([function_arguments_key, name])
-        # we use the name twice now, so that the final part of the path also has a file extension
-    )
+    return AdlsFqn.parse(invocation_output_uri(storage_root, name=name))