+++ libs/mops/src/thds/mops/pure/core/output_naming.py
@@ -3,6 +3,8 @@ import typing as ty
 
 from thds.core.stack_context import StackContext
 
+from . import types, uris
+
 PipelineFunctionUniqueKey = StackContext("Mops2PipelineFunctionUniqueKey", default="")
 FunctionArgumentsHashUniqueKey = StackContext("Mops2FunctionArgumentsHashUniqueKey", default="")
 
@@ -22,3 +24,29 @@ def pipeline_function_invocation_unique_key() -> ty.Optional[ty.Tuple[str, str]]
     if "" in pfi_key:  # if either of the elements was not supplied, we don't have anything!
         return None
     return pfi_key
+
+
+def invocation_output_uri(storage_root: uris.UriIsh = "", name: str = "") -> str:
+    """If your function only outputs a single blob, you can safely use this without
+    providing a name.  However, if you have multiple outputs from the same invocation, you
+    must provide a meaningful name for each one.
+
+    As an example:
+
+    <pipeline> <function mod/name  > <your name     > <args,kwargs hash                                   >
+    nppes/2023/thds.nppes.intake:run/<name goes here>/CoastOilAsset.IVZ9KplQKlNgxQHav0jIMUS9p4Kbn3N481e0Uvs
+    """
+    storage_root = str(storage_root or uris.ACTIVE_STORAGE_ROOT())
+    pf_fa = pipeline_function_invocation_unique_key()
+    if not pf_fa:
+        raise types.NotARunnerContext(
+            "`invocation_output_uri` must be used in a `thds.mops.pure` runner context."
+        )
+    pipeline_function_key, function_arguments_key = pf_fa
+    return uris.lookup_blob_store(storage_root).join(
+        storage_root,
+        pipeline_function_key,
+        name,
+        "--".join([function_arguments_key, name]),
+        # we use the name twice now, so that the final part of the path also has a file extension
+    )