+++ libs/mops/src/thds/mops/pure/core/uris.py
@@ -1,7 +1,9 @@
 import io
+from pathlib import Path
 from typing import Callable, Union
 
 from thds.adls import ADLS_SCHEME, AdlsFqn, AdlsRoot
+from thds.core.files import is_file_uri, to_uri
 from thds.core.stack_context import StackContext
 
 # we import the ADLS blob store here even though this is the only place in core where we 'touch'
@@ -10,18 +12,23 @@ from thds.core.stack_context import StackContext
 # In theory this could be abstracted via a registration process instead,
 # but that seems like over-engineering at this point.
 from ..adls.blob_store import AdlsBlobStore
+from .file_blob_store import FileBlobStore
 from .types import BlobStore
 
 
 def lookup_blob_store(uri: str) -> BlobStore:
     if uri.startswith(ADLS_SCHEME):
         return AdlsBlobStore()
+    if is_file_uri(uri):
+        return FileBlobStore()
     raise ValueError(f"Unsupported URI: {uri}")
 
 
 def get_root(uri: str) -> str:
     if uri.startswith(ADLS_SCHEME):
         return str(AdlsFqn.parse(uri).root())
+    if is_file_uri(uri):
+        return to_uri(Path.home())
     raise ValueError(f"Unsupported URI: {uri}")